// src/yeying/api/service/service.ts
export interface CreateServiceRequest {
  header: MessageHeader | undefined;
  body: CreateServiceRequestBody | undefined;
}

export interface CreateServiceRequestBody {
  /** 创建服务元信息入参 */
  service: ServiceMetadata | undefined;
}

export interface CreateServiceResponse {
  header: MessageHeader | undefined;
  body: CreateServiceResponseBody | undefined;
}

export interface CreateServiceResponseBody {
  status: ResponseStatus | undefined;
  service: ServiceMetadata | undefined;
}

/** detail */
export interface DetailServiceRequest {
  header: MessageHeader | undefined;
  body: DetailServiceRequestBody | undefined;
}

export interface DetailServiceRequestBody {
  /** 服务身份 did */
  did: string;
  /** 服务版本 */
  version: number;
}

export interface DetailServiceResponse {
  header: MessageHeader | undefined;
  body: DetailServiceResponseBody | undefined;
}

export interface DetailServiceResponseBody {
  status: ResponseStatus | undefined;
  service: ServiceMetadata | undefined;
}

/** search */
export interface SearchServiceRequest {
  header: MessageHeader | undefined;
  body: SearchServiceRequestBody | undefined;
}

export interface SearchServiceRequestBody {
  condition: SearchServiceCondition | undefined;
  page: RequestPage | undefined;
}

export interface SearchServiceCondition {
  /** 服务类型【参数可选】 */
  code: ServiceCodeEnum;
  /** 服务所有者【参数可选】 */
  owner: string;
  /** 服务名称【参数可选】 */
  name: string;
  /** 输入[服务所有者、服务名称]关键字搜索【参数可选】 */
  keyword: string;
}

export interface SearchServiceResponse {
  header: MessageHeader | undefined;
  body: SearchServiceResponseBody | undefined;
}

export interface SearchServiceResponseBody {
  status: ResponseStatus | undefined;
  services: ServiceMetadata[];
  page: ResponsePage | undefined;
}

/** delete */
export interface DeleteServiceRequest {
  header: MessageHeader | undefined;
  body: DeleteServiceRequestBody | undefined;
}

export interface DeleteServiceRequestBody {
  did: string;
  version: number;
}

export interface DeleteServiceResponse {
  header: MessageHeader | undefined;
  body: DeleteServiceResponseBody | undefined;
}

export interface DeleteServiceResponseBody {
  status: ResponseStatus | undefined;
}

function createBaseCreateServiceRequest(): CreateServiceRequest {
  return { header: undefined, body: undefined };
}

export const CreateServiceRequest: MessageFns<CreateServiceRequest> = {
  encode(message: CreateServiceRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateServiceRequestBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateServiceRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateServiceRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateServiceRequestBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateServiceRequest {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateServiceRequestBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateServiceRequest): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateServiceRequestBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateServiceRequest>, I>>(base?: I): CreateServiceRequest {
    return CreateServiceRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateServiceRequest>, I>>(object: I): CreateServiceRequest {
    const message = createBaseCreateServiceRequest();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateServiceRequestBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseCreateServiceRequestBody(): CreateServiceRequestBody {
  return { service: undefined };
}

export const CreateServiceRequestBody: MessageFns<CreateServiceRequestBody> = {
  encode(message: CreateServiceRequestBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.service !== undefined) {
      ServiceMetadata.encode(message.service, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateServiceRequestBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateServiceRequestBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.service = ServiceMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateServiceRequestBody {
    return { service: isSet(object.service) ? ServiceMetadata.fromJSON(object.service) : undefined };
  },

  toJSON(message: CreateServiceRequestBody): unknown {
    const obj: any = {};
    if (message.service !== undefined) {
      obj.service = ServiceMetadata.toJSON(message.service);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateServiceRequestBody>, I>>(base?: I): CreateServiceRequestBody {
    return CreateServiceRequestBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateServiceRequestBody>, I>>(object: I): CreateServiceRequestBody {
    const message = createBaseCreateServiceRequestBody();
    message.service = (object.service !== undefined && object.service !== null)
      ? ServiceMetadata.fromPartial(object.service)
      : undefined;
    return message;
  },
};

function createBaseCreateServiceResponse(): CreateServiceResponse {
  return { header: undefined, body: undefined };
}

export const CreateServiceResponse: MessageFns<CreateServiceResponse> = {
  encode(message: CreateServiceResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateServiceResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateServiceResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateServiceResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateServiceResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateServiceResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateServiceResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateServiceResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateServiceResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateServiceResponse>, I>>(base?: I): CreateServiceResponse {
    return CreateServiceResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateServiceResponse>, I>>(object: I): CreateServiceResponse {
    const message = createBaseCreateServiceResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateServiceResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};
// ...
export const SearchServiceRequest: MessageFns<SearchServiceRequest> = {
  encode(message: SearchServiceRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      SearchServiceRequestBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SearchServiceRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSearchServiceRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = SearchServiceRequestBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SearchServiceRequest {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? SearchServiceRequestBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: SearchServiceRequest): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = SearchServiceRequestBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SearchServiceRequest>, I>>(base?: I): SearchServiceRequest {
    return SearchServiceRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SearchServiceRequest>, I>>(object: I): SearchServiceRequest {
    const message = createBaseSearchServiceRequest();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? SearchServiceRequestBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};
// ...
export interface ServiceServer extends UntypedServiceImplementation {
  /** 服务创建 */
  create: handleUnaryCall<CreateServiceRequest, CreateServiceResponse>;
  /** 服务详情 */
  detail: handleUnaryCall<DetailServiceRequest, DetailServiceResponse>;
  /** 搜索服务 */
  search: handleUnaryCall<SearchServiceRequest, SearchServiceResponse>;
  /** 删除服务 */
  delete: handleUnaryCall<DeleteServiceRequest, DeleteServiceResponse>;
}
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}